Changed Byte to Integer for TINYINTs in AdqlColumn

+++ firethorn-core/src/main/java/uk/ac/roe/wfau/firethorn/meta/adql/AdqlColumn.java
@@ -259,7 +259,7 @@ extends BaseColumn<AdqlColumn>
                 case TIMESTAMP :
                     return AdqlColumn.AdqlType.DATETIME;
                 case TINYINT :
-                    return AdqlColumn.AdqlType.BYTE;
+                    return AdqlColumn.AdqlType.INTEGER;
 
 
                 case BLOB :
